OpenTween.Models にあるTwitter関連のクラスを OpenTween.SocialProtocol.Twitter に移動
