+++ OpenTween.Tests/Models/StatusTextHistoryTest.cs
@@ -19,6 +19,7 @@
 // the Free Software Foundation, Inc., 51 Franklin Street - Fifth Floor,
 // Boston, MA 02110-1301, USA.
 
+using OpenTween.SocialProtocol.Twitter;
 using Xunit;
 
 namespace OpenTween.Models
