+++ OpenTween.Tests/Models/TabInformationTest.cs
@@ -23,11 +23,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Windows.Forms;
 using OpenTween.SocialProtocol;
+using OpenTween.SocialProtocol.Twitter;
 using Xunit;
-using Xunit.Extensions;
 
 namespace OpenTween.Models
 {
