+++ OpenTween.Tests/Models/TabModelTest.cs
@@ -20,13 +20,10 @@
 // Boston, MA 02110-1301, USA.
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
+using OpenTween.SocialProtocol.Twitter;
 using Xunit;
-using Xunit.Extensions;
 
 namespace OpenTween.Models
 {
