+++ OpenTween/Api/GraphQL/CreateRetweetRequest.cs
@@ -25,7 +25,7 @@ using System;
 using System.Threading.Tasks;
 using System.Xml.XPath;
 using OpenTween.Connection;
-using OpenTween.Models;
+using OpenTween.SocialProtocol.Twitter;
 
 namespace OpenTween.Api.GraphQL
 {
