+++ OpenTween/MyCommon.cs
@@ -51,7 +51,6 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Web;
 using System.Windows.Forms;
-using OpenTween.Models;
 using OpenTween.Setting;
 using OpenTween.SocialProtocol.Twitter;
