+++ OpenTween/TimelineListViewState.cs
@@ -27,7 +27,6 @@
 #nullable enable
 
 using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
